requirements.txt
cairosvg
Pillow
tqdm
colorama
rich